Rotator Widget: Use circular distance for azimuth deadband

+++ ui/RotatorWidget.h
@@ -76,7 +76,7 @@ private:
 
     const int MAP_RESOLUTION = 1000;
     const float GLOBE_RADIUS = 100.0;
-    const int AZIMUTH_DEAD_BAND = 2;
+    const double AZIMUTH_DEAD_BAND = 2.0;
 };
 
 #endif // QLOG_UI_ROTATORWIDGET_H
